backend/pyproject.toml
[project]
name = "blogagentic"
version = "0.1.0"
description = "AI Blog Generator Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-core>=0.3.65",
    "langchain-groq>=0.3.2",
    "langgraph>=0.4.8",
    "langgraph-cli[inmem]>=0.3.1",
    "uvicorn>=0.34.3",
    "watchdog>=6.0.0",
    "python-dotenv>=1.0.1",
    "logtail-python>=0.3.0",
    "pydantic>=2.9.0",
    "pytest>=8.3.0",
    "httpx>=0.27.2",
]
